release-notes-app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
openai
azure-identity
bs4
pydantic>=2.11.5
pydantic-core>=2.23.0
cffi>=1.15.1
cryptography>=41.0.0
jiter
pytest